refactor vocab page
