+++ src/app/vocab/page.tsx
@@ -1,9 +1,9 @@
-"use client";
+'use client';
 
-import { useState, useEffect, useCallback } from "react";
-import Link from "next/link";
-import NavHeader from "@/components/NavHeader";
-import VocabList from "@/components/VocabList";
+import { useState, useEffect, useCallback } from 'react';
+import Link from 'next/link';
+import NavHeader from '@/components/NavHeader';
+import VocabList from '@/components/VocabList';
 import {
   type VocabEntry,
   type Collection,
@@ -14,289 +14,17 @@ import {
   getAllCollections,
   deleteVocabEntry,
   markVocabPushedToAnki,
-} from "@/lib/data-layer";
+} from '@/lib/data-layer';
 import {
   addBasicCard,
   addClozeCard,
   syncWordStates,
   isAnkiConnected,
   getDeckNames,
-} from "@/lib/anki";
+} from '@/lib/anki';
+import VocabStats from './components/VocabStats';
+import VocabDetailModal from './components/VocabDetailModal';
 
-// Modal for viewing/editing a vocab entry
-function VocabDetailModal({
-  entry,
-  onClose,
-  onUpdate,
-  onDelete,
-}: {
-  entry: VocabEntry;
-  onClose: () => void;
-  onUpdate: (id: string, updates: Partial<VocabEntry>) => Promise<void>;
-  onDelete: (id: string) => Promise<void>;
-}) {
-  const [isEditing, setIsEditing] = useState(false);
-  const [translation, setTranslation] = useState(entry.translation);
-  const [state, setState] = useState<WordState>(entry.state);
-  const [isSaving, setIsSaving] = useState(false);
-  const [isDeleting, setIsDeleting] = useState(false);
-
-  const handleSave = async () => {
-    setIsSaving(true);
-    try {
-      await onUpdate(entry.id, { translation, state });
-      setIsEditing(false);
-    } finally {
-      setIsSaving(false);
-    }
-  };
-
-  const handleDelete = async () => {
-    if (!confirm("Are you sure you want to delete this vocabulary entry?")) {
-      return;
-    }
-    setIsDeleting(true);
-    try {
-      await onDelete(entry.id);
-      onClose();
-    } finally {
-      setIsDeleting(false);
-    }
-  };
-
-  // State color classes for dropdown
-  const stateOptions: { value: WordState; label: string; color: string }[] = [
-    { value: "new", label: "New", color: "bg-gray-200" },
-    { value: "level1", label: "Level 1", color: "bg-blue-800" },
-    { value: "level2", label: "Level 2", color: "bg-blue-600" },
-    { value: "level3", label: "Level 3", color: "bg-blue-400" },
-    { value: "level4", label: "Level 4", color: "bg-blue-200" },
-    { value: "known", label: "Known", color: "bg-green-500" },
-    { value: "ignored", label: "Ignored", color: "bg-gray-400" },
-  ];
-
-  return (
-    <div
-      className="fixed inset-0 z-50 flex items-center justify-center bg-black/50 p-4"
-      onClick={onClose}
-    >
-      <div
-        className="w-full max-w-lg rounded-lg bg-white p-6 shadow-xl dark:bg-gray-900"
-        onClick={(e) => e.stopPropagation()}
-      >
-        {/* Header */}
-        <div className="mb-4 flex items-start justify-between">
-          <div>
-            <h2 className="text-xl font-bold text-gray-900 dark:text-gray-100">
-              {entry.text}
-            </h2>
-            <span className="text-sm text-gray-500 dark:text-gray-400">
-              {entry.type === "phrase" ? "Phrase" : "Word"}
-            </span>
-          </div>
-          <button
-            onClick={onClose}
-            className="rounded-full p-1 text-gray-400 hover:bg-gray-100 hover:text-gray-600 dark:hover:bg-gray-800 dark:hover:text-gray-300"
-          >
-            <svg
-              className="h-6 w-6"
-              fill="none"
-              viewBox="0 0 24 24"
-              stroke="currentColor"
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth={2}
-                d="M6 18L18 6M6 6l12 12"
-              />
-            </svg>
-          </button>
-        </div>
-
-        {/* Content */}
-        <div className="space-y-4">
-          {/* Translation */}
-          <div>
-            <label className="mb-1 block text-sm font-medium text-gray-700 dark:text-gray-300">
-              Translation
-            </label>
-            {isEditing ? (
-              <textarea
-                value={translation}
-                onChange={(e) => setTranslation(e.target.value)}
-                className="w-full rounded-lg border border-gray-300 px-3 py-2 text-sm focus:border-blue-500 focus:outline-none focus:ring-1 focus:ring-blue-500 dark:border-gray-600 dark:bg-gray-800 dark:text-gray-100"
-                rows={2}
-              />
-            ) : (
-              <p className="text-gray-900 dark:text-gray-100">
-                {entry.translation}
-              </p>
-            )}
-          </div>
-
-          {/* Context Sentence */}
-          <div>
-            <label className="mb-1 block text-sm font-medium text-gray-700 dark:text-gray-300">
-              Context Sentence
-            </label>
-            <p className="rounded-lg bg-gray-50 p-3 text-sm italic text-gray-700 dark:bg-gray-800 dark:text-gray-300">
-              {entry.sentence}
-            </p>
-          </div>
-
-          {/* State */}
-          <div>
-            <label className="mb-1 block text-sm font-medium text-gray-700 dark:text-gray-300">
-              Learning State
-            </label>
-            {isEditing ? (
-              <select
-                value={state}
-                onChange={(e) => setState(e.target.value as WordState)}
-                className="w-full rounded-lg border border-gray-300 px-3 py-2 text-sm focus:border-blue-500 focus:outline-none focus:ring-1 focus:ring-blue-500 dark:border-gray-600 dark:bg-gray-800 dark:text-gray-100"
-              >
-                {stateOptions.map((opt) => (
-                  <option key={opt.value} value={opt.value}>
-                    {opt.label}
-                  </option>
-                ))}
-              </select>
-            ) : (
-              <div className="flex items-center gap-2">
-                <span
-                  className={`h-3 w-3 rounded-full ${
-                    stateOptions.find((o) => o.value === entry.state)?.color
-                  }`}
-                />
-                <span className="text-gray-900 dark:text-gray-100">
-                  {stateOptions.find((o) => o.value === entry.state)?.label}
-                </span>
-              </div>
-            )}
-          </div>
-
-          {/* Metadata */}
-          <div className="grid grid-cols-2 gap-4 text-sm">
-            <div>
-              <span className="text-gray-500 dark:text-gray-400">Added: </span>
-              <span className="text-gray-900 dark:text-gray-100">
-                {new Date(entry.createdAt).toLocaleDateString("en-AU", {
-                  day: "2-digit",
-                  month: "short",
-                  year: "numeric",
-                })}
-              </span>
-            </div>
-            <div>
-              <span className="text-gray-500 dark:text-gray-400">
-                Review Count:{" "}
-              </span>
-              <span className="text-gray-900 dark:text-gray-100">
-                {entry.reviewCount}
-              </span>
-            </div>
-            {entry.pushedToAnki && (
-              <div className="col-span-2">
-                <span className="inline-flex items-center rounded-full bg-green-100 px-2 py-0.5 text-xs font-medium text-green-800 dark:bg-green-900 dark:text-green-200">
-                  Synced to Anki
-                  {entry.ankiNoteId && ` (Note #${entry.ankiNoteId})`}
-                </span>
-              </div>
-            )}
-          </div>
-        </div>
-
-        {/* Actions */}
-        <div className="mt-6 flex items-center justify-between border-t border-gray-200 pt-4 dark:border-gray-700">
-          <button
-            onClick={handleDelete}
-            disabled={isDeleting}
-            className="inline-flex items-center gap-1 rounded-lg px-3 py-2 text-sm font-medium text-red-600 hover:bg-red-50 disabled:opacity-50 dark:text-red-400 dark:hover:bg-red-900/20"
-          >
-            {isDeleting ? "Deleting..." : "Delete"}
-          </button>
-
-          <div className="flex gap-2">
-            {isEditing ? (
-              <>
-                <button
-                  onClick={() => {
-                    setIsEditing(false);
-                    setTranslation(entry.translation);
-                    setState(entry.state);
-                  }}
-                  className="rounded-lg border border-gray-300 px-4 py-2 text-sm font-medium text-gray-700 hover:bg-gray-50 dark:border-gray-600 dark:text-gray-300 dark:hover:bg-gray-800"
-                >
-                  Cancel
-                </button>
-                <button
-                  onClick={handleSave}
-                  disabled={isSaving}
-                  className="rounded-lg bg-blue-600 px-4 py-2 text-sm font-medium text-white hover:bg-blue-700 disabled:opacity-50"
-                >
-                  {isSaving ? "Saving..." : "Save"}
-                </button>
-              </>
-            ) : (
-              <button
-                onClick={() => setIsEditing(true)}
-                className="rounded-lg bg-blue-600 px-4 py-2 text-sm font-medium text-white hover:bg-blue-700"
-              >
-                Edit
-              </button>
-            )}
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-// Stats summary component
-function VocabStats({
-  stats,
-}: {
-  stats: { total: number; byState: Record<WordState, number> };
-}) {
-  const learningCount =
-    stats.byState.new +
-    stats.byState.level1 +
-    stats.byState.level2 +
-    stats.byState.level3 +
-    stats.byState.level4;
-
-  return (
-    <div className="grid grid-cols-2 gap-3 sm:grid-cols-4">
-      <div className="rounded-lg bg-gray-100 p-3 dark:bg-gray-800">
-        <div className="text-2xl font-bold text-gray-900 dark:text-gray-100">
-          {stats.total}
-        </div>
-        <div className="text-sm text-gray-600 dark:text-gray-400">
-          Total Words
-        </div>
-      </div>
-      <div className="rounded-lg bg-blue-100 p-3 dark:bg-blue-900/30">
-        <div className="text-2xl font-bold text-blue-700 dark:text-blue-400">
-          {learningCount}
-        </div>
-        <div className="text-sm text-blue-600 dark:text-blue-400">Learning</div>
-      </div>
-      <div className="rounded-lg bg-green-100 p-3 dark:bg-green-900/30">
-        <div className="text-2xl font-bold text-green-700 dark:text-green-400">
-          {stats.byState.known}
-        </div>
-        <div className="text-sm text-green-600 dark:text-green-400">Known</div>
-      </div>
-      <div className="rounded-lg bg-gray-100 p-3 dark:bg-gray-800">
-        <div className="text-2xl font-bold text-gray-500 dark:text-gray-500">
-          {stats.byState.ignored}
-        </div>
-        <div className="text-sm text-gray-500 dark:text-gray-500">Ignored</div>
-      </div>
-    </div>
-  );
-}
 
 export default function VocabPage() {
   const [entries, setEntries] = useState<VocabEntry[]>([]);
@@ -308,23 +36,22 @@ export default function VocabPage() {
   const [isLoading, setIsLoading] = useState(true);
   const [selectedEntry, setSelectedEntry] = useState<VocabEntry | null>(null);
   const [ankiConnected, setAnkiConnected] = useState<boolean | null>(null);
-  const [ankiDeck, setAnkiDeck] = useState("Afrikaans");
-  const [ankiClozeDeck, setAnkiClozeDeck] = useState("Afrikaans::Cloze");
+  const [ankiDeck, setAnkiDeck] = useState('Afrikaans');
+  const [ankiClozeDeck, setAnkiClozeDeck] = useState('Afrikaans::Cloze');
   const [notification, setNotification] = useState<{
-    type: "success" | "error";
+    type: 'success' | 'error';
     message: string;
   } | null>(null);
 
-  // Load data on mount
   useEffect(() => {
     loadData();
     checkAnkiConnection();
     // Load deck names from settings — match the keys the settings page writes
-    const savedDeck = localStorage.getItem("lector-anki-deck");
+    const savedDeck = localStorage.getItem('lector-anki-deck');
     if (savedDeck) {
       setAnkiDeck(savedDeck);
     }
-    const savedClozeDeck = localStorage.getItem("lector-anki-cloze-deck");
+    const savedClozeDeck = localStorage.getItem('lector-anki-cloze-deck');
     if (savedClozeDeck) {
       setAnkiClozeDeck(savedClozeDeck);
     }
@@ -342,8 +69,8 @@ export default function VocabPage() {
       setCollections(collectionsData);
       setStats(statsData);
     } catch (error) {
-      console.error("Failed to load data:", error);
-      showNotification("error", "Failed to load vocabulary data");
+      console.error('Failed to load data:', error);
+      showNotification('error', 'Failed to load vocabulary data');
     } finally {
       setIsLoading(false);
     }
@@ -356,14 +83,12 @@ export default function VocabPage() {
       if (connected) {
         const decks = await getDeckNames();
         // Only auto-select a deck if the user hasn't saved a preference
-        const savedDeck = localStorage.getItem("lector-anki-deck");
+        const savedDeck = localStorage.getItem('lector-anki-deck');
         if (!savedDeck) {
-          const afrikaansDeck = decks.find((d) =>
-            d.toLowerCase().includes("afrikaans")
-          );
+          const afrikaansDeck = decks.find((d) => d.toLowerCase().includes('afrikaans'));
           if (afrikaansDeck) {
             setAnkiDeck(afrikaansDeck);
-          } else if (decks.length > 0 && decks[0] !== "Default") {
+          } else if (decks.length > 0 && decks[0] !== 'Default') {
             setAnkiDeck(decks[0]);
           }
         }
@@ -373,7 +98,7 @@ export default function VocabPage() {
     }
   };
 
-  const showNotification = (type: "success" | "error", message: string) => {
+  const showNotification = (type: 'success' | 'error', message: string) => {
     setNotification({ type, message });
     setTimeout(() => setNotification(null), 5000);
   };
@@ -384,10 +109,7 @@ export default function VocabPage() {
   };
 
   // Update entry
-  const handleUpdateEntry = async (
-    id: string,
-    updates: Partial<VocabEntry>
-  ) => {
+  const handleUpdateEntry = async (id: string, updates: Partial<VocabEntry>) => {
     try {
       // Use updateVocabState for state changes, or a general update API call
       if (updates.state) {
@@ -399,10 +121,10 @@ export default function VocabPage() {
       if (selectedEntry?.id === id) {
         setSelectedEntry((prev) => (prev ? { ...prev, ...updates } : null));
       }
-      showNotification("success", "Entry updated successfully");
+      showNotification('success', 'Entry updated successfully');
     } catch (error) {
-      console.error("Failed to update entry:", error);
-      showNotification("error", "Failed to update entry");
+      console.error('Failed to update entry:', error);
+      showNotification('error', 'Failed to update entry');
       throw error;
     }
   };
@@ -412,10 +134,10 @@ export default function VocabPage() {
     try {
       await deleteVocabEntry(id);
       await loadData();
-      showNotification("success", "Entry deleted");
+      showNotification('success', 'Entry deleted');
     } catch (error) {
-      console.error("Failed to delete entry:", error);
-      showNotification("error", "Failed to delete entry");
+      console.error('Failed to delete entry:', error);
+      showNotification('error', 'Failed to delete entry');
       throw error;
     }
   };
@@ -428,20 +150,15 @@ export default function VocabPage() {
     async (ids: string[], cardType: 'basic' | 'cloze') => {
       if (!ankiConnected) {
         showNotification(
-          "error",
-          "Anki is not connected. Make sure Anki is running with AnkiConnect."
+          'error',
+          'Anki is not connected. Make sure Anki is running with AnkiConnect.',
         );
         return;
       }
 
-      const entriesToExport = entries.filter(
-        (e) => ids.includes(e.id) && !e.pushedToAnki
-      );
+      const entriesToExport = entries.filter((e) => ids.includes(e.id) && !e.pushedToAnki);
       if (entriesToExport.length === 0) {
-        showNotification(
-          "error",
-          "All selected entries have already been synced to Anki."
-        );
+        showNotification('error', 'All selected entries have already been synced to Anki.');
         return;
       }
 
@@ -459,7 +176,7 @@ export default function VocabPage() {
             entry.sentence,
             entry.text,
             entry.translation,
-            entry.translation // word meaning — using translation for now
+            entry.translation, // word meaning — using translation for now
           );
           await markVocabPushedToAnki(entry.id, noteId);
           successCount++;
@@ -473,52 +190,46 @@ export default function VocabPage() {
 
       if (errorCount === 0) {
         showNotification(
-          "success",
-          `Exported ${successCount} ${cardLabel} card${successCount === 1 ? '' : 's'} to "${targetDeck}"`
+          'success',
+          `Exported ${successCount} ${cardLabel} card${successCount === 1 ? '' : 's'} to "${targetDeck}"`,
         );
       } else {
         showNotification(
-          "error",
-          `Exported ${successCount} ${cardLabel} cards, ${errorCount} failed`
+          'error',
+          `Exported ${successCount} ${cardLabel} cards, ${errorCount} failed`,
         );
       }
     },
-    [entries, ankiConnected, ankiDeck, ankiClozeDeck]
+    [entries, ankiConnected, ankiDeck, ankiClozeDeck],
   );
 
   // Mark selected entries as known
-  const handleMarkAsKnown = useCallback(
-    async (ids: string[]) => {
-      try {
-        for (const id of ids) {
-          await updateVocabState(id, "known");
-        }
-        await loadData();
-        showNotification(
-          "success",
-          `Marked ${ids.length} entries as known`
-        );
-      } catch (error) {
-        console.error("Failed to mark as known:", error);
-        showNotification("error", "Failed to mark entries as known");
+  const handleMarkAsKnown = useCallback(async (ids: string[]) => {
+    try {
+      for (const id of ids) {
+        await updateVocabState(id, 'known');
       }
-    },
-    []
-  );
+      await loadData();
+      showNotification('success', `Marked ${ids.length} entries as known`);
+    } catch (error) {
+      console.error('Failed to mark as known:', error);
+      showNotification('error', 'Failed to mark entries as known');
+    }
+  }, []);
 
   // Sync with Anki to update mastery levels
   const handleSyncWithAnki = useCallback(async () => {
     if (!ankiConnected) {
       showNotification(
-        "error",
-        "Anki is not connected. Make sure Anki is running with AnkiConnect."
+        'error',
+        'Anki is not connected. Make sure Anki is running with AnkiConnect.',
       );
       return;
     }
 
     try {
       // Get deck name from settings
-      const deckName = localStorage.getItem("lector-anki-deck") || ankiDeck;
+      const deckName = localStorage.getItem('lector-anki-deck') || ankiDeck;
       console.log(`Syncing with Anki deck: "${deckName}"`);
       const wordStates = await syncWordStates(deckName);
       console.log(`Found ${wordStates.size} words in Anki`);
@@ -542,7 +253,7 @@ export default function VocabPage() {
         const ankiData = wordStates.get(entry.text.toLowerCase());
         if (ankiData) {
           matchedCount++;
-          if (entry.state === "ignored") continue;
+          if (entry.state === 'ignored') continue;
           // New/relearning cards (interval < 1 day) carry no signal yet.
           if (ankiData.interval < 1) continue;
 
@@ -554,15 +265,15 @@ export default function VocabPage() {
           // 31+ days: known
           let newState: WordState;
           if (ankiData.interval >= 31) {
-            newState = "known";
+            newState = 'known';
           } else if (ankiData.interval >= 15) {
-            newState = "level4";
+            newState = 'level4';
           } else if (ankiData.interval >= 8) {
-            newState = "level3";
+            newState = 'level3';
           } else if (ankiData.interval >= 2) {
-            newState = "level2";
+            newState = 'level2';
           } else {
-            newState = "level1";
+            newState = 'level1';
           }
 
           if (stateRank[newState] > stateRank[entry.state]) {
@@ -574,20 +285,20 @@ export default function VocabPage() {
 
       await loadData();
       showNotification(
-        "success",
-        `Found ${wordStates.size} cards in "${deckName}". Matched ${matchedCount} vocab entries, updated ${updatedCount}.`
+        'success',
+        `Found ${wordStates.size} cards in "${deckName}". Matched ${matchedCount} vocab entries, updated ${updatedCount}.`,
       );
     } catch (error) {
-      console.error("Failed to sync with Anki:", error);
-      showNotification("error", "Failed to sync with Anki");
+      console.error('Failed to sync with Anki:', error);
+      showNotification('error', 'Failed to sync with Anki');
     }
   }, [entries, ankiConnected, ankiDeck]);
 
   return (
-    <div className="min-h-screen bg-gray-50 dark:bg-gray-950 pt-[var(--mobile-topbar-h)] sm:pt-0 sm:ml-56">
+    <div className="min-h-screen bg-gray-50 pt-[var(--mobile-topbar-h)] sm:ml-56 sm:pt-0 dark:bg-gray-950">
       <NavHeader />
       {/* Header — mobile only, desktop uses sidebar */}
-      <header className="sm:hidden border-b border-gray-200 bg-white dark:border-gray-800 dark:bg-gray-900">
+      <header className="border-b border-gray-200 bg-white sm:hidden dark:border-gray-800 dark:bg-gray-900">
         <div className="mx-auto max-w-7xl px-4 py-4 sm:px-6 lg:px-8">
           <div className="flex items-center justify-between">
             <div className="flex items-center gap-4">
@@ -595,12 +306,7 @@ export default function VocabPage() {
                 href="/"
                 className="text-gray-500 hover:text-gray-700 dark:text-gray-400 dark:hover:text-gray-200"
               >
-                <svg
-                  className="h-6 w-6"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke="currentColor"
-                >
+                <svg className="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor">
                   <path
                     strokeLinecap="round"
                     strokeLinejoin="round"
@@ -609,17 +315,13 @@ export default function VocabPage() {
                   />
                 </svg>
               </Link>
-              <h1 className="text-2xl font-bold text-gray-900 dark:text-gray-100">
-                Vocabulary
-              </h1>
+              <h1 className="text-2xl font-bold text-gray-900 dark:text-gray-100">Vocabulary</h1>
             </div>
 
             {/* Anki connection status */}
             <div className="flex items-center gap-2">
               {ankiConnected === null ? (
-                <span className="text-sm text-gray-500">
-                  Checking Anki connection...
-                </span>
+                <span className="text-sm text-gray-500">Checking Anki connection...</span>
               ) : ankiConnected ? (
                 <span className="inline-flex items-center gap-1.5 rounded-full bg-green-100 px-3 py-1 text-sm font-medium text-green-800 dark:bg-green-900 dark:text-green-200">
                   <span className="h-2 w-2 rounded-full bg-green-500" />
@@ -639,20 +341,14 @@ export default function VocabPage() {
       {/* Notification */}
       {notification && (
         <div
-          className={`fixed right-4 top-4 z-50 rounded-lg px-4 py-3 shadow-lg ${
-            notification.type === "success"
-              ? "bg-green-100 text-green-800 dark:bg-green-900 dark:text-green-200"
-              : "bg-red-100 text-red-800 dark:bg-red-900 dark:text-red-200"
-          }`}
+          className={`fixed top-4 right-4 z-50 rounded-lg px-4 py-3 shadow-lg ${notification.type === 'success'
+            ? 'bg-green-100 text-green-800 dark:bg-green-900 dark:text-green-200'
+            : 'bg-red-100 text-red-800 dark:bg-red-900 dark:text-red-200'
+            }`}
         >
           <div className="flex items-center gap-2">
-            {notification.type === "success" ? (
-              <svg
-                className="h-5 w-5"
-                fill="none"
-                viewBox="0 0 24 24"
-                stroke="currentColor"
-              >
+            {notification.type === 'success' ? (
+              <svg className="h-5 w-5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
                 <path
                   strokeLinecap="round"
                   strokeLinejoin="round"
@@ -661,12 +357,7 @@ export default function VocabPage() {
                 />
               </svg>
             ) : (
-              <svg
-                className="h-5 w-5"
-                fill="none"
-                viewBox="0 0 24 24"
-                stroke="currentColor"
-              >
+              <svg className="h-5 w-5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
                 <path
                   strokeLinecap="round"
                   strokeLinejoin="round"
@@ -676,16 +367,8 @@ export default function VocabPage() {
               </svg>
             )}
             <span>{notification.message}</span>
-            <button
-              onClick={() => setNotification(null)}
-              className="ml-2 hover:opacity-70"
-            >
-              <svg
-                className="h-4 w-4"
-                fill="none"
-                viewBox="0 0 24 24"
-                stroke="currentColor"
-              >
+            <button onClick={() => setNotification(null)} className="ml-2 hover:opacity-70">
+              <svg className="h-4 w-4" fill="none" viewBox="0 0 24 24" stroke="currentColor">
                 <path
                   strokeLinecap="round"
                   strokeLinejoin="round"
@@ -709,7 +392,7 @@ export default function VocabPage() {
 
         {/* Vocabulary List - exclude ignored words */}
         <VocabList
-          entries={entries.filter(e => e.state !== 'ignored')}
+          entries={entries.filter((e) => e.state !== 'ignored')}
           collections={collections}
           onEntryClick={handleEntryClick}
           onExportToAnki={handleExportToAnki}
